requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
pillow==10.4.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tzlocal==5.2
